requirements.txt
atproto>=0.0.54
anthropic>=0.40.0
python-dotenv>=1.0.0
google-cloud-firestore>=2.26.0
functions-framework>=3.0.0
